put each legacy codecs in their own package
This allows us to strictly segregate each codec version (to make sure
that we don't accidentally mix codec versions), while still letting
tests access each unitary codecs (which using inner `private classes`
would have prevented). Relevant tests only need to be moved to the
appropriate package.

```
wire
 |
 `-- ChannelCodecs
 |
 `-- legacy
 |      |
 |      `-- legacy0
 |      |      |
 |      |      `-- LegacyChannelCodecs0
 |      |
 |      `-- legacy1
 |      |      |
 |      |      `-- LegacyChannelCodecs1
 |      |
 |      `-- ChannelTypes
 |
 `-- others
```
